Move eigenfunctions.py out of the root folder (issue #43)
