+++ geometric_kernels/kernels/geometric_kernels.py
@@ -5,10 +5,10 @@ Implementation of geometric kernels on several spaces
 import lab as B
 import numpy as np
 
-from geometric_kernels.eigenfunctions import Eigenfunctions
 from geometric_kernels.kernels.base import BaseGeometricKernel
 from geometric_kernels.lab_extras import from_numpy, logspace, trapz
 from geometric_kernels.spaces.base import DiscreteSpectrumSpace
+from geometric_kernels.spaces.eigenfunctions import Eigenfunctions
 from geometric_kernels.spaces.hyperbolic import Hyperbolic
 from geometric_kernels.utils.utils import Optional
