+++ geometric_kernels/spaces/eigenfunctions.py
