+++ geometric_kernels/spaces/mesh.py
@@ -9,9 +9,9 @@ import potpourri3d as pp3d
 import robust_laplacian
 import scipy.sparse.linalg as sla
 
-from geometric_kernels.eigenfunctions import Eigenfunctions
 from geometric_kernels.lab_extras import from_numpy, take_along_axis
 from geometric_kernels.spaces import DiscreteSpectrumSpace
+from geometric_kernels.spaces.eigenfunctions import Eigenfunctions
 from geometric_kernels.utils.utils import Optional
