+++ tests/spaces/test_circle.py
@@ -6,10 +6,10 @@ import torch
 from opt_einsum import contract as einsum
 from plum import Tuple
 
-from geometric_kernels.eigenfunctions import EigenfunctionWithAdditionTheorem
 from geometric_kernels.kernels.geometric_kernels import MaternKarhunenLoeveKernel
 from geometric_kernels.lab_extras import from_numpy
 from geometric_kernels.spaces.circle import Circle, SinCosEigenfunctions
+from geometric_kernels.spaces.eigenfunctions import EigenfunctionWithAdditionTheorem
 from geometric_kernels.utils.utils import chain
